+++ src/helpers/constants/Player.js
@@ -3,6 +3,7 @@ exports.playerAttributes = [
     { attribute: "seen", source: "seer", remainingPhases: 1 },
     { attribute: "eaten", source: "werewolves", remainingPhases: 1 },
     { attribute: "eaten", source: "big-bad-wolf", remainingPhases: 1 },
+    { attribute: "infected", source: "vile-father-of-wolves", remainingPhases: 1 },
     { attribute: "drank-life-potion", source: "witch", remainingPhases: 1 },
     { attribute: "drank-death-potion", source: "witch", remainingPhases: 1 },
     { attribute: "protected", source: "guard", remainingPhases: 1 },