+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### 🚀 New features
 
+* [#73](https://github.com/antoinezanardi/werewolves-assistant-api/issues/73) - Vile Father of Wolves Role implemented.
 * [#86](https://github.com/antoinezanardi/werewolves-assistant-api/issues/86) - Extend player limit to 40.
 * [#87](https://github.com/antoinezanardi/werewolves-assistant-api/issues/87) - Add isRevealed to role structure.
 * [#91](https://github.com/antoinezanardi/werewolves-assistant-api/issues/91) - Game option for no sheriff.